+++ src/main/java/com/ubt/hospitalmanagement/dtos/requests/ReviewRequest.java
@@ -7,6 +7,6 @@ import lombok.Data;
 public class ReviewRequest {
 
     private String review;
-    private long doctorId;
+    private Integer doctorId;
 
 }
