+++ src/main/java/com/ubt/hospitalmanagement/repositories/UserRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String userName);
 
     Optional<User> findByUuid(String uuid);
-    Optional<User> findByIdAndRolesContaining(Long id, String role);
+    Optional<User> findByIdAndRolesContaining(Integer id, String role);
 
     Page<User> findByRolesContaining(String role, Pageable pageable);
 }
